Separated demodulator
