+++ fobrob.c
@@ -17,6 +17,7 @@
 #include "protocol.h"
 #include "runningavg.h"
 #include "filter.h"
+#include "demodulator.h"
 
 void handlePacket(unsigned char* packet)
 {
@@ -87,208 +88,6 @@ void printHelp(char *s)
 	fprintf(stderr, "\t-d, --debug <int>            : debug stuff. Look in the code if you're interested\n");
 }
 
-typedef struct 
-{
-	/* Bit demod */
-	int state;
-	int preambleGood;
-	int lastDemodBit;
-	int ending;
-	int minPreamble;
-	int manchPos;
-	int bitpos;
-	uint8_t byte;
-	int bytePos;
-	uint8_t packet[10];
-	int retry;
-
-	/* Sample demod */
-	int lastBit;
-	int sampleAt;
-	int sampleNum;
-	int samplesPerBit;
-	runningAvgContext midPointCtx;
-	runningAvgContext bitAvgCtx;
-} DemodContext;
-
-#define DEMOD_PREAMBLE		0
-#define DEMOD_WAITZEROES	1
-#define DEMOD_BITBANG		2
-#define DEMOD_WAITRETRY		3
-void demodBit(DemodContext* demodCtx, int bit)
-{
-	switch (demodCtx->state)
-	{
-		case DEMOD_PREAMBLE: // Wait for preamble
-			/* Check for a series of alternating bits */
-			if (demodCtx->lastDemodBit != bit)
-			{
-				demodCtx->preambleGood++;
-			}
-			else
-			{
-				demodCtx->preambleGood = 0;
-			}
-
-			/* Do we have enough of a preamble to be confident that it's the start of a packet ? */
-			if (demodCtx->preambleGood >= demodCtx->minPreamble)
-			{
-				demodCtx->state = DEMOD_WAITZEROES;
-				demodCtx->preambleGood = 0;
-				demodCtx->ending = 0;
-			}
-			break;
-		case DEMOD_WAITZEROES: // wait for en of preamble: 4 zeroes in a row
-			if (bit == 0)
-			{
-				demodCtx->ending++;
-				/* Get ready for the next state if we have our 4 zeroes */
-				if (demodCtx->ending == 4)
-				{
-					fprintf(stderr, "Got preamble!\n");
-					demodCtx->ending = 0;
-					demodCtx->manchPos = 0;
-					demodCtx->byte = 0;
-					demodCtx->bitpos = 0;
-					demodCtx->bytePos = 0;
-					demodCtx->retry = 1;
-					demodCtx->state = DEMOD_BITBANG;
-				}
-			}
-			else
-			{
-				demodCtx->ending = 0;
-				if (demodCtx->ending > 1)
-				{
-					/* We just had more than 1 zero followed by a one: not a valid preamble: reset state */
-					demodCtx->ending = 0;
-					demodCtx->state = DEMOD_PREAMBLE;
-				}
-			}
-
-			break;
-		case DEMOD_BITBANG: // bit bang
-			if (demodCtx->manchPos & 1)
-			{       
-				/* Check for manchester encoding errors */
-				if (bit == demodCtx->lastDemodBit)
-				{
-//					fprintf(stderr, "WARNING: Manchester encoding error: %d and %d at manchPos %d\n", bit, demodCtx->lastDemodBit, demodCtx->manchPos);
-
-					if (!demodCtx->retry)
-					{
-						/* Reset state */
-						demodCtx->state = DEMOD_PREAMBLE;
-						demodCtx->preambleGood = 0;
-					}
-				}
-			}
-			else
-			{
-				/* This is a data bit */
-				demodCtx->byte <<= 1;
-				demodCtx->byte |= bit;
-
-				demodCtx->bitpos++;
-
-				/* Do we have a full byte? */
-				if (demodCtx->bitpos == 8)
-				{       
-					/* Add to buffer */
-					demodCtx->packet[demodCtx->bytePos++] = demodCtx->byte;
-					demodCtx->byte = 0;
-					demodCtx->bitpos = 0;
-
-					/* Do we have a full packet? */
-					if (demodCtx->bytePos == sizeof(demodCtx->packet))
-					{
-						if (isValidPacket(demodCtx->packet) == -1)
-						{
-							fprintf(stderr, "Invalid packet\n");
-							if (demodCtx->retry == 0)
-							{
-								/* No more retries: reset state */
-								demodCtx->state = DEMOD_PREAMBLE;
-							}
-							else
-							{
-								/* Try the next transmission? */
-								demodCtx->state = DEMOD_WAITRETRY;
-							}
-						}
-						else
-						{
-							handlePacket(demodCtx->packet);
-							demodCtx->preambleGood = 0;
-							demodCtx->state = DEMOD_PREAMBLE;
-						}
-
-						demodCtx->retry--;
-					}
-				}
-			}
-
-			demodCtx->manchPos++;
-			break;
-
-		case DEMOD_WAITRETRY: // Wait 5 bits for the next iransmission
-			demodCtx->manchPos++;
-			if (demodCtx->manchPos == 164)
-			{
-				demodCtx->manchPos = 0;
-				demodCtx->byte = 0;
-				demodCtx->bitpos = 0;
-				demodCtx->bytePos = 0;
-
-				demodCtx->state = DEMOD_BITBANG;
-			}
-			break;
-
-	}
-
-	demodCtx->lastDemodBit = bit;
-}
-
-void demodSample(DemodContext* demodCtx, double magnitude)
-{
-	/* And get a running average of one bitlength of that */
-	double avgSample = runningAvg(&demodCtx->bitAvgCtx, magnitude);
-
-	/* Get a mid point to compare levels against */
-	double midPoint = runningAvg(&demodCtx->midPointCtx, avgSample);
-
-	/* Get the value of the actual bit */
-	int bit = avgSample > midPoint;
-
-	/* Check for zero-crossing */
-	if (demodCtx->lastBit != bit)
-	{
-		// Align the sampleAt variable here
-		demodCtx->sampleAt = demodCtx->sampleNum + demodCtx->samplesPerBit / 2.0;
-	}
-
-	/* Is this the middle of our sample ? */
-	if (demodCtx->sampleNum >= demodCtx->sampleAt)
-	{
-		demodBit(demodCtx, bit);
-
-		/* Time the next sample */
-		demodCtx->sampleAt += demodCtx->samplesPerBit;
-	}
-
-	demodCtx->lastBit = bit;
-	demodCtx->sampleNum++;
-}
-
-void demodInit(DemodContext* demodCtx, int samplesPerBit, int minPreambleBits)
-{
-	memset(demodCtx, 0, sizeof(DemodContext));
-	demodCtx->minPreamble = minPreambleBits;
-	demodCtx->samplesPerBit = samplesPerBit;
-	runningAvgInit(&demodCtx->bitAvgCtx, samplesPerBit);
-	runningAvgInit(&demodCtx->midPointCtx, samplesPerBit * 8); // XXX: 8 is a guess. Significantly longer than any period without a zero-crossing, shorter than the preamble
-}
-
 int main(int argc, char **argv)
 {
 	double freq = 433.92e+6;
@@ -305,7 +104,7 @@ int main(int argc, char **argv)
 	SampleFilter_init(&lpfq1);
 
 	DemodContext demodCtx;
-	demodInit(&demodCtx, samplesPerBit, minPreambleBits);
+	demodInit(&demodCtx, samplesPerBit, minPreambleBits, &handlePacket);
 
 	int decimator = 0;
